services/ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
openai==1.51.0
anthropic==0.39.0
sentence-transformers==3.1.0
numpy==1.26.4
supabase==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0